backend/pyproject.toml
[project]
name = "cimiento"
version = "0.1.0"
description = "Copiloto local de anteproyecto residencial"
requires-python = ">=3.11"
dependencies = [
    # Core
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    # Optimización
    "ortools>=9.11.0",
    # Geometría y BIM
    "shapely>=2.0.6",
    "ifcopenshell>=0.8.0",
    "ezdxf>=1.3.0",
    "trimesh>=4.5.0",
    "openpyxl>=3.1.5",
    "numpy>=2.1.0",
    # LLM y agentes
    "langgraph>=0.2.45",
    "langchain-ollama>=0.2.0",
    "ollama>=0.4.0",
    # RAG
    "qdrant-client>=1.12.0",
    "sentence-transformers>=3.2.0",
    # Base de datos
    "sqlalchemy>=2.0.36",
    "asyncpg>=0.30.0",
    "aiosqlite>=0.20.0",
    "alembic>=1.14.0",
    # Utilidades
    "httpx>=0.27.0",
    "structlog>=24.4.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
    "ipykernel>=6.29.0",
    "jupyter>=1.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
